+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * renamed to Maven Accio
 * added support to load projects recursively
+* removed dependencies out argument in Maven.loadPom methods - use MavenProject.getArtifacts() instead
 * Maven is AutoClosable now, close() invokes container.dispose() to fix memory leak
 * loadPom now defaults to resolve dependencies and process plugins (previously, plugins were never processed)
 * hide internals: removed Maven method getRepositorySystem(), getRepositorySession(), remoteLegacyRepositories() and remoteResolveRepositories()